+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Image editor: fix image URLs not updated for HTML syntax in markdown
 * Image editor: view original image in preview dialog
 * Image editor: close dialog on save if opend via edit image button
+* Markdown editor: support HTML attrs for code blocks
 * `cleanupfiles` command: delete unreferenced files from storage and database
 * `backup` command: allow saving backups directly to S3
 * Improve backup performance